requirements.txt
openai>=1.45.0
python-dotenv>=1.0.1
rich>=13.7.1